Refactor test util
